+++ examples/vite-react-router/src/components/Navigation.tsx
@@ -1,5 +1,9 @@
-import { Navigation as MidasNavigation, NavigationItem } from '@midas-ds/layout'
-import { House, FileText } from 'lucide-react'
+import {
+  Navigation as MidasNavigation,
+  NavigationItem,
+  NavigationSubMenu,
+} from '@midas-ds/layout'
+import { House, FileText, Plus, Save } from 'lucide-react'
 import { NavigationLink } from './NavigationLink'
 
 export const Navigation = () => (
@@ -19,6 +23,24 @@ export const Navigation = () => (
       >
         Applications
       </NavigationLink>
+      <NavigationSubMenu>
+        <NavigationItem>
+          <NavigationLink
+            to='/applications/new'
+            icon={<Plus />}
+          >
+            New
+          </NavigationLink>
+        </NavigationItem>
+        <NavigationItem>
+          <NavigationLink
+            to='/applications/drafts'
+            icon={<Save />}
+          >
+            Drafts
+          </NavigationLink>
+        </NavigationItem>
+      </NavigationSubMenu>
     </NavigationItem>
   </MidasNavigation>
 )