+++ packages/components/src/checkbox/CheckboxGroup.stories.tsx
@@ -104,3 +104,13 @@ export const Invalid: Story = {
     isInvalid: true,
   },
 }
+
+export const WithHelpPopover: Story = {
+  args: {
+    popover: {
+      children:
+        'An assistive text that helps the user understand the field better.',
+      'aria-label': 'Mer information',
+    },
+  },
+}