+++ packages/components/src/time-field/TimeField.stories.tsx
@@ -76,3 +76,14 @@ export const HourCycle12: Story = {
     defaultValue: new Time(14, 30),
   },
 }
+
+export const WithHelpPopover: Story = {
+  args: {
+    ...Primary.args,
+    popover: {
+      children:
+        'An assistive text that helps the user understand the field better.',
+      'aria-label': 'Mer information',
+    },
+  },
+}